Adaptive layout improved for landscape and tablet.

+++ app/src/main/java/com/openclassrooms/realestatemanager/view/model_ui/PropertyUi.kt
@@ -7,7 +7,7 @@ object EmptyProperty : PropertyWrapper()
 data class PropertyUi (
 
         val type: String = "HOUSE",
-        val price: Float,
+        val price: Int,
         val area: Float,
         val roomNbr: Int,
         val description: String,
